+++ src/main/java/com/sponus/sponusbe/domain/announcement/controller/AnnouncementController.java
@@ -13,11 +13,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.sponus.sponusbe.auth.annotation.AuthOrganization;
-import com.sponus.sponusbe.domain.announcement.dto.AnnouncementCreateRequest;
-import com.sponus.sponusbe.domain.announcement.dto.AnnouncementCreateResponse;
-import com.sponus.sponusbe.domain.announcement.dto.AnnouncementResponse;
-import com.sponus.sponusbe.domain.announcement.dto.AnnouncementUpdateRequest;
-import com.sponus.sponusbe.domain.announcement.dto.AnnouncementUpdateResponse;
+import com.sponus.sponusbe.domain.announcement.dto.request.AnnouncementCreateRequest;
+import com.sponus.sponusbe.domain.announcement.dto.request.AnnouncementUpdateRequest;
+import com.sponus.sponusbe.domain.announcement.dto.response.AnnouncementCreateResponse;
+import com.sponus.sponusbe.domain.announcement.dto.response.AnnouncementResponse;
+import com.sponus.sponusbe.domain.announcement.dto.response.AnnouncementUpdateResponse;
 import com.sponus.sponusbe.domain.announcement.service.AnnouncementQueryService;
 import com.sponus.sponusbe.domain.announcement.service.AnnouncementService;
 import com.sponus.sponusbe.domain.organization.entity.Organization;
