--- src/main/java/com/sponus/sponusbe/domain/announcement/dto/AnnouncementBriefResponse.java
@@ -1,4 +1,0 @@
-package com.sponus.sponusbe.domain.announcement.dto;
-
-public record AnnouncementBriefResponse() {
-}
