+++ src/main/java/com/sponus/sponusbe/domain/announcement/dto/request/AnnouncementCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sponus.sponusbe.domain.announcement.dto;
+package com.sponus.sponusbe.domain.announcement.dto.request;
 
 import com.sponus.sponusbe.domain.announcement.entity.Announcement;
 import com.sponus.sponusbe.domain.announcement.entity.enums.AnnouncementCategory;
@@ -8,10 +8,8 @@ import com.sponus.sponusbe.domain.organization.entity.Organization;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
 
-@Builder
-public record AnnouncementUpdateRequest(
+public record AnnouncementCreateRequest(
 	@NotBlank(message = "[ERROR] 타이틀 입력은 필수 입니다.")
 	String title,
 	@NotNull(message = "[ERROR] 유형 입력은 필수 입니다.")
