+++ src/main/java/com/sponus/sponusbe/domain/announcement/dto/response/AnnouncementBriefResponse.java
@@ -1,0 +1,4 @@
+package com.sponus.sponusbe.domain.announcement.dto.response;
+
+public record AnnouncementBriefResponse() {
+}
